+++ graph-server/src/renderer/render-png.js
@@ -72,12 +72,6 @@ async function getPNG({pageContent, viewportSize, componentOptions}) {
         // Use an arbitrary width that will guarantee desktop-like rendering.
         page.setViewport(viewportSize);
 
-        // Set the page content from the pre-built bundle.
-        // await Promise.all([
-        //     page.addStyleTag({content: bundles.styles}),
-        //     page.addScriptTag({content: bundles.script})
-        // ]);
-
         // Log browser console messages
         page.on('console', msg => console.log('[CONSOLE LOG]', msg.text()));
 
@@ -86,7 +80,7 @@ async function getPNG({pageContent, viewportSize, componentOptions}) {
 
         // If the page should have a comparison series, wait for it. Otherwise,
         // wait for the current year series. Fall back to the no-data-message
-        const chartSel = componentOptions.compare ? '.ts-compare' : '.ts-current';
+        const chartSel = componentOptions.compare ? '#ts-compare-group' : '#ts-current-group';
         await page.waitForSelector(`${chartSel}, #no-data-message`);
 
         // Get a snapshot of either the graph or the no data message, if the