+++ CHANGELOG.md
@@ -4,15 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/usgs/waterdataui/compare/waterdataui-0.26.0...master)
+## [Unreleased](https://github.com/usgs/waterdataui/compare/waterdataui-0.27.0...master)
+## Fixed
+- Fixed bug that made compare time series disappear when the brush was adjusted by adding
+offsets to time series state in order to adjust x scale.
+
+## [0.27.0](https://github.com/usgs/waterdataui/compare/waterdataui-0.26.0...waterdataui-0.27.0) - 2020-3-11
 ### Changed
 - All IV time series are now available for drawing on the hydrograph not just those in the last 7 days.
 - Added brush handle fill.
+- Updated to latest wdfn-viz 1.4.0 and implemented the new build guidance.
 
 ### Added
 - The state of the UI for the hydrograph is now preserved in the hash portion of the url and state 
 can be restored by pasting the url into a browser window.
+- Added upstream basin to NLDI map on the monitoring location pages.
 - Added both hover and a cursor slider tooltip capability to the DV graph.
+- Added width integer parameter to the graph-server which allows the user to return the width of the image.
 
 ## [0.26.0](https://github.com/usgs/waterdataui/compare/waterdataui-0.25.0...waterdataui-0.26.0) - 2020-02-12
 ### Added